Relocated stateless fragment helpers out of AbstractScoringTask
* Moved HasTopNFragmentMatch + GetTop6FragmentMzs (+cache) to new OspreySharp.Core.FragmentMath; CountTopNFragmentOverlap + TopNFragmentMzs to new OspreySharp.Scoring.FragmentOverlap (verbatim; split by dependency since the overlap counter uses ScoringMath)
* Pure relocation: Stellar + Astral cross-impl 1e-9 bit-parity PASS, 345/347 tests, no perf regression
* Flagged the dead TheoreticalIsotopeEnvelope in place (zero callers; not relocated)

See ai/todos/active/TODO-20260530_ospreysharp_relocate_domain_helpers.md
